Trait hotline::gfx::Device

pub trait Device: Sized + Any {
    type SwapChain: SwapChain<Self>;
    type CmdBuf: CmdBuf<Self>;
    type Buffer: Buffer<Self>;
    type Shader: Shader<Self>;
    type RenderPipeline: RenderPipeline<Self>;
    type Texture: Texture<Self>;
    type ReadBackRequest: ReadBackRequest<Self>;
    type RenderPass: RenderPass<Self>;
    type Heap: Heap<Self>;
    type ComputePipeline: ComputePipeline<Self>;

    fn create(info: &DeviceInfo) -> Self;

    fn create_heap(&self, info: &HeapInfo) -> Self::Heap;

    fn create_swap_chain(
        &mut self, 
        info: &SwapChainInfo, 
        window: &Window
    ) -> Self::SwapChain;

    fn create_cmd_buf(&self, num_buffers: u32) -> Self::CmdBuf;

    fn create_shader<T: Sized>(
        &self, 
        info: &ShaderInfo, 
        src: &[T]
    ) -> Result<Self::Shader, Error>;

    fn create_buffer<T: Sized>(
        &mut self, 
        info: &BufferInfo, 
        data: Option<&[T]>
    ) -> Result<Self::Buffer, Error>;

    fn create_texture<T: Sized>(
        &mut self, 
        info: &TextureInfo, 
        data: Option<&[T]>
    ) -> Result<Self::Texture, Error>;

    fn create_render_pipeline(
        &self, 
        info: &RenderPipelineInfo<'_, Self>
    ) -> Result<Self::RenderPipeline, Error>;

    fn create_render_pass(
        &self, 
        info: &RenderPassInfo<Self>
    ) -> Result<Self::RenderPass, Error>;

    fn create_compute_pipeline(
        &self, 
        info: &ComputePipelineInfo<Self>
    ) -> Result<Self::ComputePipeline, Error>;

    fn execute(&self, cmd: &Self::CmdBuf);

    fn get_shader_heap(&self) -> &Self::Heap;

    fn get_adapter_info(&self) -> &AdapterInfo;

}

A GPU device is used to create GPU resources, the device also contains a single a single command queue to which all command buffers will submitted and executed each frame.

Associated Types

type SwapChain: SwapChain<Self>

type CmdBuf: CmdBuf<Self>

type Buffer: Buffer<Self>

type Shader: Shader<Self>

type RenderPipeline: RenderPipeline<Self>

type Texture: Texture<Self>

type ReadBackRequest: ReadBackRequest<Self>

type RenderPass: RenderPass<Self>

type Heap: Heap<Self>

type ComputePipeline: ComputePipeline<Self>

Required methods

fn create(info: &DeviceInfo) -> Self

fn create_heap(&self, info: &HeapInfo) -> Self::Heap

fn create_swap_chain(
&mut self,
info: &SwapChainInfo,
window: &Window
) -> Self::SwapChain

fn create_cmd_buf(&self, num_buffers: u32) -> Self::CmdBuf

fn create_shader<T: Sized>(
&self,
info: &ShaderInfo,
src: &[T]
) -> Result<Self::Shader, Error>

fn create_buffer<T: Sized>(
&mut self,
info: &BufferInfo,
data: Option<&[T]>
) -> Result<Self::Buffer, Error>

fn create_texture<T: Sized>(
&mut self,
info: &TextureInfo,
data: Option<&[T]>
) -> Result<Self::Texture, Error>

fn create_render_pipeline(
&self,
info: &RenderPipelineInfo<'_, Self>
) -> Result<Self::RenderPipeline, Error>

fn create_render_pass(
&self,
info: &RenderPassInfo<Self>
) -> Result<Self::RenderPass, Error>

fn create_compute_pipeline(
&self,
info: &ComputePipelineInfo<Self>
) -> Result<Self::ComputePipeline, Error>

fn execute(&self, cmd: &Self::CmdBuf)

fn get_shader_heap(&self) -> &Self::Heap

fn get_adapter_info(&self) -> &AdapterInfo