Trait hotline::os::App

pub trait App: 'static + Any + Sized {
    type Window: Window<Self>;
    type NativeHandle: NativeHandle<Self>;

    fn create(info: AppInfo) -> Self;

    fn create_window(&self, info: WindowInfo<Self>) -> Self::Window;

    fn run(&mut self) -> bool;

    fn get_mouse_pos(&self) -> Point<i32>;

    fn get_mouse_wheel(&self) -> f32;

    fn get_mouse_hwheel(&self) -> f32;

    fn get_mouse_buttons(&self) -> [bool; 5];

    fn get_utf16_input(&self) -> Vec<u16>;

    fn get_keys_down(&self) -> [bool; 256];

    fn is_sys_key_down(&self, key: SysKey) -> bool;

    fn get_key_code(key: Key) -> i32;

    fn enumerate_display_monitors() -> Vec<MonitorInfo>;

    fn set_cursor(&self, cursor: &Cursor);

}

An interface which all platforms need to implement for general operating system calls

Associated Types

type Window: Window<Self>

type NativeHandle: NativeHandle<Self>

Required methods

fn create(info: AppInfo) -> Self

Create an application instance

fn create_window(&self, info: WindowInfo<Self>) -> Self::Window

Create a new operating system window

fn run(&mut self) -> bool

Call to update windows and os state each frame, when false is returned the app has been requested to close

fn get_mouse_pos(&self) -> Point<i32>

Retuns the mouse in screen coordinates

fn get_mouse_wheel(&self) -> f32

Retuns the mouse vertical wheel position

fn get_mouse_hwheel(&self) -> f32

Retuns the mouse horizontal wheel positions

fn get_mouse_buttons(&self) -> [bool; 5]

Retuns the mouse button states, up or down

fn get_utf16_input(&self) -> Vec<u16>

Returns a vector of utf-16 characters that have been input since the last frame

fn get_keys_down(&self) -> [bool; 256]

Returns an array of bools containing 0-256 keys down (true) or up (false)

fn is_sys_key_down(&self, key: SysKey) -> bool

Returns true if the sys key is down and false if the key is up

fn get_key_code(key: Key) -> i32

Get os system virtual key code from Key

fn enumerate_display_monitors() -> Vec<MonitorInfo>

Enumerate all display monitors

fn set_cursor(&self, cursor: &Cursor)

Sets the mouse cursor