Trait hotline::os::Window

pub trait Window<A: App>: Any + Sized {

    fn bring_to_front(&self);

    fn show(&self, show: bool, activate: bool);

    fn update(&mut self);

    fn close(&mut self);

    fn update_style(&mut self, flags: WindowStyleFlags, rect: Rect<i32>);

    fn is_focused(&self) -> bool;

    fn is_minimised(&self) -> bool;

    fn set_focused(&self);

    fn is_mouse_hovered(&self) -> bool;

    fn set_title(&self, title: String);

    fn set_pos(&self, pos: Point<i32>);

    fn set_size(&self, size: Size<i32>);

    fn get_pos(&self) -> Point<i32>;

    fn get_viewport_rect(&self) -> Rect<i32>;

    fn get_size(&self) -> Size<i32>;

    fn get_mouse_client_pos(&self, mouse_pos: Point<i32>) -> Point<i32>;

    fn get_dpi_scale(&self) -> f32;

    fn get_native_handle(&self) -> A::NativeHandle;

    fn get_events(&self) -> WindowEventFlags;

    fn clear_events(&mut self);

    fn as_ptr(&self) -> *const Self;

    fn as_mut_ptr(&mut self) -> *mut Self;

}

An instance of an operating system window

Required methods

fn bring_to_front(&self)

Bring window to front and draw ontop of all others

fn show(&self, show: bool, activate: bool)

Show window, specify true to show window or false to hide

fn update(&mut self)

Must be called each frame to handle resizes

fn close(&mut self)

Close the window

fn update_style(&mut self, flags: WindowStyleFlags, rect: Rect<i32>)

Change the windows style

fn is_focused(&self) -> bool

Returns true if the window is focused

fn is_minimised(&self) -> bool

Returns true if the window is minimised

fn set_focused(&self)

Sets focus to this window

fn is_mouse_hovered(&self) -> bool

Returns true if the mouse if hovering this window

fn set_title(&self, title: String)

Set the window display title that appears on the title bar

fn set_pos(&self, pos: Point<i32>)

Set window position in screen space

fn set_size(&self, size: Size<i32>)

Set window size in screen coordinates

fn get_pos(&self) -> Point<i32>

Returns the screen position for the top-left corner of the window

fn get_viewport_rect(&self) -> Rect<i32>

Returns a gfx friendly full window rect to use as gfx::Viewport or gfx::Scissor

fn get_size(&self) -> Size<i32>

Returns the screen position for the top-left corner of the window

fn get_mouse_client_pos(&self, mouse_pos: Point<i32>) -> Point<i32>

Return mouse position in relative coordinates from the top left corner of the window

fn get_dpi_scale(&self) -> f32

Return the dpi scale for the current monitor the window is on

fn get_native_handle(&self) -> A::NativeHandle

Gets the internal native handle

fn get_events(&self) -> WindowEventFlags

Gets window events tracked from os update, to handle events inside external systems

fn clear_events(&mut self)

Clears events after they have been responded to

fn as_ptr(&self) -> *const Self

Const pointer

fn as_mut_ptr(&mut self) -> *mut Self

Mut pointer