Trait hotline::gfx::CmdBuf

pub trait CmdBuf<D: Device> {

    fn reset(&mut self, swap_chain: &D::SwapChain);

    fn close(&mut self, swap_chain: &D::SwapChain);

    fn begin_render_pass(&self, render_pass: &mut D::RenderPass);

    fn end_render_pass(&self);

    fn begin_event(&self, colour: u32, name: &str);

    fn end_event(&self);

    fn transition_barrier(&mut self, barrier: &TransitionBarrier<D>);

    fn set_viewport(&self, viewport: &Viewport);

    fn set_scissor_rect(&self, scissor_rect: &ScissorRect);

    fn set_index_buffer(&self, buffer: &D::Buffer);

    fn set_vertex_buffer(&self, buffer: &D::Buffer, slot: u32);

    fn set_render_pipeline(&self, pipeline: &D::RenderPipeline);

    fn set_compute_pipeline(&self, pipeline: &D::ComputePipeline);

    fn set_compute_heap(&self, slot: u32, heap: &D::Heap);

    fn set_render_heap(&self, slot: u32, heap: &D::Heap);

    fn set_marker(&self, colour: u32, name: &str);

    fn push_constants<T: Sized>(
        &self, 
        slot: u32, 
        num_values: u32, 
        dest_offset: u32, 
        data: &[T]
    );

    fn draw_instanced(
        &self, 
        vertex_count: u32, 
        instance_count: u32, 
        start_vertex: u32, 
        start_instance: u32
    );

    fn draw_indexed_instanced(
        &self, 
        index_count: u32, 
        instance_count: u32, 
        start_index: u32, 
        base_vertex: i32, 
        start_instance: u32
    );

    fn dispatch(&self, group_count: Size3, thread_count: Size3);

    fn read_back_backbuffer(
        &mut self, 
        swap_chain: &D::SwapChain
    ) -> D::ReadBackRequest;

}

Responsible for buffering graphics commands. Internally it will contain a platform specific command list for each buffer in the associated swap chain. At the start of each frame reset must be called with an associated swap chain to internally switch which buffer we are writing to. At the end of each frame close must be called and finally the CmdBuf can be passed to Device::execute to be processed on the GPU.

Required methods

fn reset(&mut self, swap_chain: &D::SwapChain)

fn close(&mut self, swap_chain: &D::SwapChain)

fn begin_render_pass(&self, render_pass: &mut D::RenderPass)

fn end_render_pass(&self)

fn begin_event(&self, colour: u32, name: &str)

fn end_event(&self)

fn transition_barrier(&mut self, barrier: &TransitionBarrier<D>)

fn set_viewport(&self, viewport: &Viewport)

fn set_scissor_rect(&self, scissor_rect: &ScissorRect)

fn set_index_buffer(&self, buffer: &D::Buffer)

fn set_vertex_buffer(&self, buffer: &D::Buffer, slot: u32)

fn set_render_pipeline(&self, pipeline: &D::RenderPipeline)

fn set_compute_pipeline(&self, pipeline: &D::ComputePipeline)

fn set_compute_heap(&self, slot: u32, heap: &D::Heap)

fn set_render_heap(&self, slot: u32, heap: &D::Heap)

fn set_marker(&self, colour: u32, name: &str)

fn push_constants<T: Sized>(
&self,
slot: u32,
num_values: u32,
dest_offset: u32,
data: &[T]
)

fn draw_instanced(
&self,
vertex_count: u32,
instance_count: u32,
start_vertex: u32,
start_instance: u32
)

fn draw_indexed_instanced(
&self,
index_count: u32,
instance_count: u32,
start_index: u32,
base_vertex: i32,
start_instance: u32
)

fn dispatch(&self, group_count: Size3, thread_count: Size3)

Thread count is required for metal, in hlsl it is specified in the shader

fn read_back_backbuffer(
&mut self,
swap_chain: &D::SwapChain
) -> D::ReadBackRequest