Trait hotline::os::App

pub trait App: 'static + Any + Sized {
    type Window: Window<Self>;
    fn create(info: AppInfo) -> Self;

    fn create_window(&self, info: WindowInfo) -> Self::Window;

    fn run(&self) -> bool;
}

An interface which all platforms need to implement for general operating system calls

Associated Types

type Window: Window<Self>

Required methods

fn create(info: AppInfo) -> Self

Create an application instance

fn create_window(&self, info: WindowInfo) -> Self::Window

Create a new operating system window

fn run(&self) -> bool

Call to update windows and os state each frame, when false is returned the app has been requested to close